Stopping for the weekend. All JSON objects successfully created. Still needs testing.
I'll update the readme file in a bit.  Nice to have would be pass CSVReader a fully qualified path, currently hard-coded. Should update SQL response sets to prepared statements to prevent SQL-injection. Needs through testing and more functional comments. Be nice to expose the database population procedure in a URI, shouldn't be a lot of effort. A maven build/.pom file would greatly simplify client installation. I will document the steps in readme file as long as a description of the service calls.